Move input_pipeline to utils.py
